22860

See merge request akita/akita-erp!66

+++ resources/assets/js/controller/customers-list-vl.js
@@ -68,11 +68,29 @@ var ctrCustomersListVl = new Vue({
             }
         },
         deleteSupplier: function (id){
-            if (confirm(messages["MSG06001"])) {
-                customers_service.deleteSupplier(id).then((response) => {
+            customers_service.checkIsExist(id).then((response) => {
+                if (!response.success) {
+                    alert(response.msg);
                     this.getItems(1);
-                });
-            }
+                    return false;
+                } else {
+                    if (confirm(messages["MSG06001"])) {
+                        customers_service.deleteSupplier(id).then((response) => {
+                            this.getItems(1);
+                        });
+                    }
+                }
+            });
+        },
+        checkIsExist: function (id) {
+            customers_service.checkIsExist(id).then((response) => {
+                if (!response.success) {
+                    alert(response.msg);
+                    this.getItems(1);
+                } else {
+                    window.location.href = 'edit/' + id;
+                }
+            });
         }
     },
     mounted () {
